refactor(login): modularize form logic and optimize rendering performance
- Extract `LoginForm` component to isolate input state and validation logic
- Move error styles to `login-form.css` and remove obsolete `login-page.css`
- Memoize `LoginHeader` and `CityLocationItem` to eliminate unnecessary repaints
- Implement `useCallback` for handlers to ensure referential stability
- Resolve flamegraph bottleneck where keystrokes triggered full page reconciliation
